feat : 임시저장 버튼으로도 isDrafting이 감지되도록 수정
